Move GUI launcher function into qt subpackage
