+++ geoAssembler/tests/conftest.py
@@ -65,7 +65,7 @@ def save_geo():
 @pytest.fixture()
 def calib(gui_app):
     """Create the calibration gui"""
-    from geoAssembler.qt.pyqt import QtMainWidget
+    from geoAssembler.qt.app import QtMainWidget
 
     main_gui = QtMainWidget(gui_app)
     yield main_gui
